meta-ivi: Use corresponding N-0.2 commit on branch
860f0ce dbus: remove "not essential patches for common-api"

meta-ivi was pointing to N-0.2 tag but the tagged commit is not on any
published branch.  When doing a submodule update the tag was not fetched,
which could lead some builds to fail unless an explicit "git fetch
--tags" was performed.

An equivalent state exists on the 12.0 branch, so we use that commit for
the moment.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>