Add scripts/ci-build.sh
This helper script takes care of a lot of details to simplify setup of
GDP builds in CI.  Instead of the steps being defined in the CI tool,
they are defined here.  Over time many details have crept into the
build definitions that were intially simple.  Those details are not
always applicable for building locally by hand, but the script could be
useful locally *if* the user understands what it does.