Use only git:// URLs for submodules
We're currently having some strange network issues in CI when updating
submodules.  Coinciding with this, there were *confirmed* failures in
accessing GitHub using HTTP while other protocols worked fine - see
this ticket:

[GDP-726] HTTP connection reset by peer on go-agent1 and go-agent2
(This was supposed to be an IT-infrastructure ticket, not a GDP ticket)

We had a mix of http://, https:// and git://.  There's nothing in that
configuration that "should" not work, but this seemed like a good
opportunity to update it to be consistent, while testing if this might
have any positive effect on the failures.

In the past we have favored HTTPS URLs for easier usage through various
firewalls, but since I found the majority of submodules used git:// URLs
before this change, it's clear that making them all use git:// is as
good/bad as any other choice.

Everyone must run "git submodule sync" for a change like this but that is
included in init.sh so it will be transparent to most.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>